+++ src/App.js
@@ -1,25 +1,13 @@
-import React, { Context } from "react";
+import React from "react";
 
-import { MiniContext } from "./storage/context";
-import { AppRouter } from "./components/router/AppRouter";
-import { Header } from "./components/shared/Header";
-
-import "./App.css";
+import { ContextData, MiniContextProvider } from "storage/context";
+import { AppLayout } from "./AppLayout";
 
 function App() {
-    const urlPath = document.location.pathname;
-
     return (
-        <MiniContext.Provider value={MiniContext.profile}>
-            <div className="App">
-                {/* Begin: Menu - Fixed part of the app */}
-                <Header />
-                {/* End: Menu - Fixed part of the app */}
-                {/* Begin: Moving part of the app */}
-                <AppRouter routePath={urlPath} />
-                {/* End: Moving part of the app */}
-            </div>
-        </MiniContext.Provider>
+        <MiniContextProvider value={ContextData}>
+            <AppLayout />
+        </MiniContextProvider>
     );
 }