+++ src/storage/context/index.js
@@ -1,21 +1,26 @@
 import React from "react";
+import { UserProfile } from "models/UserProfile";
 
-const dataStructure = {
-    profile: {
-        name: "",
-        mail: "",
-        id: ""
-    },
-    setProfile: function ({ name, mail, id }) {
-        debugger;
+const ContextData = {
+    profile: new UserProfile(),
+    setProfile: async function ({ name, mail, id } = new UserProfile()) {
         this.profile.id = id;
         this.profile.name = name;
         this.profile.mail = mail;
+    },
+    setProfileName: async function (name = "") {
+        this.profile.name = name;
+    },
+    selectProfileName: async function () {
+        return this.profile.name;
     }
 };
 
-const MiniContext = React.createContext(dataStructure);
+const MiniContext = React.createContext(ContextData);
 
 MiniContext.displayName = "MiniContext";
 
-export { MiniContext };
+const MiniContextProvider = MiniContext.Provider;
+const MiniContextConsumer = MiniContext.Consumer;
+
+export { ContextData, MiniContext, MiniContextProvider, MiniContextConsumer };